backend/requirements.txt
# Core FastAPI
fastapi
uvicorn>=0.24.0
pydantic
pydantic-settings
python-multipart

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Environment and utils
python-dotenv>=1.0.0
aiohttp
numpy
redis
slowapi

# Authentication
pyjwt>=2.8.0
requests
python-jose[cryptography]>=3.3.0
cryptography>=39.0.1
svix>=1.4.0

# AI and LangChain
openai
google-generativeai
langchain>=0.0.267
langserve>=0.0.30
# Specify minimal versions to reduce dependencies
langchain-community>=0.0.16
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# Vector store and embeddings
faiss-cpu
# Use a slimmer version of sentence-transformers without full models
sentence-transformers==2.2.2

# Document processing
arxiv
pypdf

# Document generation
reportlab>=4.0.8
markdown2>=2.4.10
jinja2>=3.1.2